third_party/WebKit/LayoutTests/fast/forms/textarea-resize-below-min-size-set-expected.txt

Test for resizing the Textarea below the minimum size set.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById("textInputID").style.width is "194px"
PASS document.getElementById("textInputID").style.height; is "194px"
PASS document.getElementById("textInputID").style.width is "114px"
PASS document.getElementById("textInputID").style.height; is "84px"
PASS document.getElementById("textInputID").style.width is "74px"
PASS document.getElementById("textInputID").style.height; is "74px"
PASS successfullyParsed is true

TEST COMPLETE
